+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Updated types to allow for select menus in modals, as well as using a `LabelComponent` as using action rows in modals are now deprecated.
 ## [6.5.0] - 2025-07-17
 ### Added:
 - Added `collectibles` to users